moved obtainFrameValueLst to data.py and added data import to app.py
