+++ src/utils.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pandas as pd
 from typing import List, Tuple
 
 
@@ -48,26 +47,6 @@ def derivative(values: List[float]) -> List[float]:
     dy = [values[i+1] - values[i] for i in range(len(values)-1)]
     dy.append(dy[-1])
     return dy
-
-
-def obtainFrameValueLst(df: pd.DataFrame) -> Tuple[List[int], List[float]]:
-    """ Function to retrieve the columns of a pandas dataframe corresponding to the frames and the pixel value of each
-    frame.
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        Pandas dataframe containing the frames column and the pixel value column as a minimum.
-
-    Returns
-    -------
-    Tuple[List[int], List[float]]
-        Tuple containing two lists where the first lists corresponds to the column containing the subsequent indexes of
-        frames and the second list contains the pixel intensity value of each frame.
-    """
-    frames = df['Frame']
-    values = df['Pixel value']
-    return frames.tolist(), values.tolist()
 
 
 def getMax(frames: List[int], values: List[float], threshold: int) -> Tuple[List[int], List[float]]:
